requirements.txt
yfinance>=0.2.37
pandas>=2.2.0
numpy>=1.26.0
lightgbm>=4.1.0
scikit-learn>=1.4.0
tqdm>=4.66.0
pyarrow>=15.0.0
joblib>=1.3.0
pyyaml>=6.0.0
matplotlib>=3.8.0
pytest>=8.0.0
xlsxwriter>=3.2.0
tradingview_ta>=3.3.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.24.0
webdriver-manager>=4.0.2